move "list" classes to proper package
