+++ jaxygen-api-sample/src/main/java/org/jaxygen/netserviceapisample/business/dto/lists/ArrayListExampleResponseDTO.java
@@ -13,20 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jaxygen.netserviceapisample.business.dto;
+package org.jaxygen.netserviceapisample.business.dto.lists;
 
 /**
  *
  * @author jknast
  */
-public class GenericArrayListBase {
-  private GenericList<UserDTO> genericListStrings = new GenericList<>();
- 
-  public GenericList<UserDTO> getGenericListStrings() {
-    return genericListStrings;
-  }
-
-  public void setGenericListStrings(GenericList<UserDTO> genericListStrings) {
-    this.genericListStrings = genericListStrings;
-  }
+public class ArrayListExampleResponseDTO extends ArrayListExampleBase{
+  
 }
