+++ jaxygen-api-sample/src/main/java/org/jaxygen/netserviceapisample/business/dto/lists/ArrayListOfObjectsRequestDTO.java
@@ -13,14 +13,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jaxygen.netserviceapisample.business.dto;
+package org.jaxygen.netserviceapisample.business.dto.lists;
 
 import java.util.ArrayList;
+import org.jaxygen.netserviceapisample.business.dto.UserDTO;
+import org.jaxygen.netserviceapisample.business.dto.UserDTO;
 
 /**
  *
  * @author jknast
  */
-public class GenericList<T> extends ArrayList<T>{
-  
+public class ArrayListOfObjectsRequestDTO {
+    private ArrayList<UserDTO> list = new ArrayList<>();
+
+    public ArrayList<UserDTO> getList() {
+        return list;
+    }
+
+    public void setList(ArrayList<UserDTO> list) {
+        this.list = list;
+    }
 }
