+++ jaxygen-api-sample/src/main/java/org/jaxygen/netserviceapisample/business/dto/lists/ArrayListWithArrayListRequestDTO.java
@@ -13,12 +13,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jaxygen.netserviceapisample.business.dto;
+package org.jaxygen.netserviceapisample.business.dto.lists;
+
+import java.util.ArrayList;
 
 /**
  *
  * @author jknast
  */
-public class ArrayListExampleResponseDTO extends ArrayListExampleBase{
-  
+public class ArrayListWithArrayListRequestDTO {
+    private ArrayList<UserDTOList> listOfList = new ArrayList<>();
+
+    public ArrayList<UserDTOList> getListOfList() {
+        return listOfList;
+    }
+
+    public void setListOfList(ArrayList<UserDTOList> listOfList) {
+        this.listOfList = listOfList;
+    }
+
+    
+
 }
